feat: Move RemoteButtonViewModel to shared usecase module (Phase 27)
Move DefaultRemoteButtonViewModel + interface from androidApp to
usecase/src/commonMain/. Add androidx.lifecycle:lifecycle-viewmodel
KMP dependency. Add allowedPrefixes to ImportBoundaryCheckTask for
KMP-compatible androidx libraries.

- Replace java.util.Date with Long (currentTimeMillis) in createButtonAckToken
- Remove @Inject annotation (DI wiring stays in AppComponent)
- Add allowedPrefixes support to import boundary checker
