+++ AndroidGarage/androidApp/src/main/java/com/chriscartland/garage/ui/HomeContent.kt
@@ -54,7 +54,7 @@ import com.chriscartland.garage.domain.model.RequestStatus
 import com.chriscartland.garage.door.DoorViewModel
 import com.chriscartland.garage.permissions.notificationJustificationText
 import com.chriscartland.garage.permissions.rememberNotificationPermissionState
-import com.chriscartland.garage.remotebutton.RemoteButtonViewModel
+import com.chriscartland.garage.usecase.RemoteButtonViewModel
 import com.google.accompanist.permissions.ExperimentalPermissionsApi
 import com.google.accompanist.permissions.PermissionState
 import com.google.accompanist.permissions.PermissionStatus
